Refactor models, update workflows, and improve database handling
